+++ src/Profile/Shopware6/Converter/LanguageConverter.php
@@ -8,6 +8,9 @@
 namespace SwagMigrationAssistant\Profile\Shopware6\Converter;
 
 use Shopware\Core\Framework\Log\Package;
+use Shopware\Core\Framework\Uuid\Uuid;
+use Shopware\Core\System\Language\LanguageEntity;
+use Shopware\Core\System\Locale\LocaleEntity;
 use SwagMigrationAssistant\Migration\Converter\ConvertStruct;
 use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
@@ -23,9 +26,9 @@ class LanguageConverter extends ShopwareConverter
             && $this->getDataSetEntity($migrationContext) === LanguageDataSet::getEntity();
     }
 
-    public function convertData(array $data): ConvertStruct
+    protected function convertData(array $data): ConvertStruct
     {
-        $converted = $data;
+        $converted = $this->checkDataForDefaultLanguage($data);
         $languageId = $this->mappingService->getLanguageUuid(
             $this->connectionId,
             $data['locale']['code'],
@@ -60,5 +63,30 @@ class LanguageConverter extends ShopwareConverter
         }
 
         return new ConvertStruct($converted, $data, $this->mainMapping['id'] ?? null);
+    }
+
+    /**
+     * @param array<string, mixed> $data
+     *
+     * @return array<string, mixed>
+     */
+    private function checkDataForDefaultLanguage(array $data): array
+    {
+        $defaultLanguage = $this->mappingService->getDefaultLanguage($this->context);
+        if (!$defaultLanguage instanceof LanguageEntity) {
+            return $data;
+        }
+
+        $defaultLocale = $defaultLanguage->getLocale();
+        if (!$defaultLocale instanceof LocaleEntity) {
+            return $data;
+        }
+
+        if ($data['id'] === $defaultLanguage->getId() && $data['locale']['code'] !== $defaultLocale->getCode()) {
+            // The ID the of given language will newly generated, because the current default language should not be overwritten
+            $data['id'] = Uuid::randomHex();
+        }
+
+        return $data;
     }
 }